requirements.txt
openai
replicate
pillow
requests
mediapipe
opencv-python
numpy
matplotlib
insightface==0.7.3
onnxruntime
onnxruntime-gpu 